+++ .gitignore
@@ -1,3 +1,4 @@
+/data/piper.gresource
 /piper.egg-info
 /build
 __pycache__
